Converted client player position packet
